adding support for preemptible GCE instances
This change adds a `--gce-preemptible` option that will flag the instance as a preemptible instance and disable auto-migrate and auto-restart. Users can choose to use this for instances where the user doesn't care if they go away and thus pay less for the instance.

Also fixes a few rubocop issues from latest chefstyle.